requirements.txt
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
numpy==1.26.4
packaging==25.0
pandas==2.1.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.6.0
